requirements.txt
Flask
Flask-CORS
python-docx
pdfplumber
